+++ sqlbuilder_insert.go
@@ -67,8 +67,8 @@ func (ib *InsertBuilder) SetMap(m map[string]any, opts ...BuilderOption) *Insert
 	}
 
 	for n, v := range m {
-		if bo.ToSnake != nil {
-			sn := bo.ToSnake(n)
+		if bo.ToName != nil {
+			sn := bo.ToName(n)
 			if sn != n {
 				delete(m, n)
 				m[sn] = v