+++ sqlbuilder_option.go
@@ -3,13 +3,13 @@ package sqle
 type BuilderOption func(opts *BuilderOptions)
 
 type BuilderOptions struct {
-	ToSnake func(string) string // db column name converter
+	ToName  func(string) string // db column naming convert method
 	Columns []string            // db column filter
 }
 
-func WithToSnake(fn func(string) string) BuilderOption {
+func WithToName(fn func(string) string) BuilderOption {
 	return func(opts *BuilderOptions) {
-		opts.ToSnake = fn
+		opts.ToName = fn
 	}
 }